+++ src/Arborescence.Models/IndexedSetPolicy.cs
@@ -1,4 +1,4 @@
-﻿namespace Arborescence.Traversal
+﻿namespace Arborescence.Models
 {
 #pragma warning disable CA1815 // Override equals and operator equals on value types
     /// <summary>
@@ -9,7 +9,7 @@
         /// <inheritdoc/>
         public bool Contains(byte[] items, int item)
         {
-            if (items == null || (uint)item >= (uint)items.Length)
+            if (items is null || (uint)item >= (uint)items.Length)
                 return false;
 
             return items[item] != 0;
@@ -18,7 +18,7 @@
         /// <inheritdoc/>
         public void Add(byte[] items, int item)
         {
-            if (items == null || (uint)item >= (uint)items.Length)
+            if (items is null || (uint)item >= (uint)items.Length)
                 return;
 
             items[item] = 1;