+++ tests/Arborescence.Traversal.Tests/QueueGenericSearchEnumerateVerticesTest.cs
@@ -5,6 +5,7 @@
     using System.Collections.Generic;
     using System.Diagnostics;
     using Misnomer;
+    using Models;
     using Traversal;
     using Xunit;
     using EdgeEnumerator = ArrayPrefixEnumerator<Endpoints>;
@@ -13,18 +14,12 @@
 
     public class QueueGenericSearchEnumerateVerticesTest
     {
-        public QueueGenericSearchEnumerateVerticesTest()
-        {
-            InstantBfs = default;
-            GenericSearch = default;
-        }
-
         private InstantBfs<Graph, int, Endpoints, EdgeEnumerator, byte[], GraphPolicy, IndexedColorMapPolicy>
-            InstantBfs { get; }
+            InstantBfs { get; } = default;
 
         private GenericSearch<
                 Graph, int, Endpoints, EdgeEnumerator, Queue<int>, byte[], GraphPolicy, QueuePolicy, IndexedSetPolicy>
-            GenericSearch { get; }
+            GenericSearch { get; } = default;
 
         private void EnumerateVerticesCore(Graph graph, bool multipleSource)
         {
@@ -101,7 +96,6 @@
             return result;
         }
 
-#pragma warning disable CA1707 // Identifiers should not contain underscores
         [Theory]
         [ClassData(typeof(MutableSimpleGraphCollection))]
         internal void EnumerateVertices_SingleSource(GraphParameter<Graph> p)
@@ -115,6 +109,5 @@
         {
             EnumerateVerticesCore(p.Graph, true);
         }
-#pragma warning restore CA1707 // Identifiers should not contain underscores
     }
 }